+++ src/main/java/com/tomas/appstates/SticksAppState.java
@@ -14,8 +14,8 @@ import com.jme3.light.LightList;
 import com.jme3.math.Vector3f;
 import com.jme3.scene.Node;
 import com.jme3.scene.Spatial;
+import com.tomas.controls.KinectHandControl;
 import com.tomas.kinect.Kinect;
-import com.tomas.kinect.control.KinectHandControl;
 import com.tomas.properties.CollisionGroups;
 import com.tomas.properties.Hand;
 import com.tomas.properties.StickData;
