+++ src/main/java/com/tomas/utils/Velocity.java
@@ -1,4 +1,4 @@
-package com.tomas.kinect.control;
+package com.tomas.utils;
 
 import com.jme3.math.Vector3f;
